Added newlines and fixed compiler warnings.

+++ OPExtensionKit/Source/Foundation/NSArray+Opetopic.m
@@ -80,7 +80,7 @@
 -(id) anyObject {
 	if ([self count] == 0)
 		return nil;
-	return [self objectAtIndex:arc4random_uniform(self.count)];
+	return [self objectAtIndex:arc4random_uniform((u_int32_t)self.count)];
 }
 
 -(id) anyObjectExcept:(id)object {
